Bug 1710933 - Save LaunchWithApplication downloads into downloads folder for windows and unix r=mtigley
Differential Revision: https://phabricator.services.mozilla.com/D115997

+++ toolkit/components/downloads/DownloadCore.jsm
@@ -598,7 +598,10 @@ Download.prototype = {
           new FileUtils.File(this.target.path)
         );
       } else if (
-        Services.prefs.getBoolPref("browser.helperApps.deleteTempFileOnExit")
+        Services.prefs.getBoolPref("browser.helperApps.deleteTempFileOnExit") &&
+        !Services.prefs.getBoolPref(
+          "browser.download.improvements_to_download_panel"
+        )
       ) {
         gExternalAppLauncher.deleteTemporaryFileOnExit(
           new FileUtils.File(this.target.path)
